+++ RELEASE_NOTES.md
@@ -2,6 +2,14 @@
 
 # Release 1.15.0 2020-06-02
 
+Issue #14398 - Handle nulls for wildcards in QARTOD records; add QARTOD code tests
+- Update conda environment/requirements
+- Wrote tests for QARTOD code
+- Added handling for nulls as wildcards in QARTOD record fields
+- Added hash, equals, print methods to QartodTestRecord object
+
+Issue #14398 - QARTOD proof of concept
+
 Issue #14278 - Use netcdf_name attribute to rename parameters in output files
 
 Issue #13674 - Exclude coordinates attribute from timestamp parameters